test: chunk bmt and inclusion proof
